Add support for xcassets
In order to load images or colors from the included ElegantCalendar.xcassets it is needed to specify on creation
which bundle should be used. So creating a color looks like `Color("some name", bundle: .bundle`. Creating an
image looks like `Image("some name", in: .bundle, with: nil)`.

The xcassets file needs to be inside the sources folder.