Add support for defining a background color for a day via the datasource
MonthlyCalendarDataSource and YearlyCalendarDataSource get a new method
calendar(backgroundColorForDate) -> Color?. The default implementation returns nil. This
will render the day with the theme color. If the datasource returns a color this is used instead.

The yearly datasource added support for ranges. If the datasource returns a color for a given
date this is uesd. If this date is at the start or end of a range a rectangle with rounded corners
left or right respectively with the given color is drawn.