Change month layout and possible options
Change the month layout so that it is more compact in order to give the accessory view more room.
Add the possibility to display coloured dots beneath the day. This can be used to show that there is
something attached to this day. Therefore add calendar(auxDotsColorsForDate) to the datasource.
Use a Capsule instead of a Circle to show the day’s color.